\name{anapuce-package}
\alias{anapuce-package}
\alias{anapuce}
\docType{package}
\title{
Tools for microarray data analysis
}
\description{
This package contains functions for normalisation,differentially analysis of microarray data and others functions implementing recent methods developed by the Statistic and Genom Team from UMR 518 AgroParisTech/INRA Appl. Math. Comput. Sc.}
\details{
\tabular{ll}{
Package: \tab anapuce\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2010-08-17\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
INRA, J. Aubert <julie.aubert@agroparistech.fr>}
\references{
\url{http://rwiki.sciviews.org/doku.php?id=packages:cran:anapuce}
\url{http://www.agroparistech.fr/mia/doku.php?id=productions:logiciels}
}
\keyword{ package }
\seealso{
\code{\link[anapuce]{normalisation}},\code{\link[anapuce]{DiffAnalysis}},
\code{\link[anapuce]{DiffAnalysis.unpaired}},\code{\link[anapuce]{LBI}},
\code{\link[anapuce]{LocalFDR}}
}
