\name{fonctionMobile}
\alias{fonctionMobile}
\title{ fonctionMobile }
\description{
  fonctionMobile
}
\usage{
fonctionMobile(x, FUN, lag)
}
\arguments{
  \item{x   }{  }
  \item{FUN }{  }
  \item{lag }{  }
}
\details{
}
\value{
 }
\author{ J. Aubert }

\seealso{ }

\keyword{ internal }
