\name{lissageMM}
\alias{lissageMM}
\title{ Mean Average Smoothing }
\description{ Internal function used in LocalFDR corresponding to mean average smoothing
}
\usage{
lissageMM(x, indice, mm)
}
\arguments{
  \item{x       }{ }
  \item{indice  }{ }
  \item{mm      }{ }
}
\author{ J. Aubert }

\seealso{ \code{\link{LocalFDR}}. }


\keyword{ internal }
