\name{anapuce-package}
\alias{anapuce-package}
\alias{anapuce}
\docType{package}
\title{
Tools for microarray data analysis
}
\description{
This package contains functions for normalisation,differentially analysis of microarray data and others functions implementing recent methods developed by the Statistic and Genom Team from UMR 518 AgroParisTech/INRA Appl. Math. Comput. Sc.}
\details{
\tabular{ll}{
Package: \tab anapuce\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2008-10-02\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
INRA, J. Aubert <julie.aubert@agroparistech.fr>}
\references{
\url{http://wiki.r-project.org/rwiki/doku.php?id=packages:cran:anapuce}
\url{http://www.agroparistech.fr/mia/}
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-anapuce]{normalisation}},\code{\link[<pkg>:<pkg>-anapuce]{DiffAnalysis}},
\code{\link[<pkg>:<pkg>-anapuce]{DiffAnalysis.unpaired}},\code{\link[<pkg>:<pkg>-anapuce]{LBI}},
\code{\link[<pkg>:<pkg>-anapuce]{LocalFDR}}
}
